riak-deadmanshand, to facilitate full riak restart from wm_cluster code
This standalone script is intended to be started separately, possibly
from a systemd unit-file. It will check, every 3 sec, for presence
of a file named RESTART_RIAK in /run/riak, and execute
`riak stop; riak start` if that file exists.

The idea is to enable a full restart of riak initiated by an operator
action taken remotely in riak_control (i.e., over http), in a situation
when `riak restart` is insufficient (importantly, restarting riak this way
will not cause cuttlefish to read and process riak.conf and
advanced.config).

This idea becomes especially useful in combination with remote editing
and pushing advanced.config and riak.conf, implemented in riak_control,
allowing operators entirely remotely to configure riak nodes and do
a rolling restart subsequently, all without a need to have ssh and
root access to riak boxes.