Use standard relx
In standard relx, USE_NODETOOL=true needs to be set get stuff on stdout.  this may be resolved in the future using -fetch_stdout  on erl_call, but not in present relx.

Previously this was resolved by modifying relx.  Now this adds USE_NODETOOL=true when we make a rpc call.

There is some circular stuff going on here. A riak admin call, will call the riak script (generated by relx) which will call the riak_admin script which will call lib.sh for a rpc() function call, which will call the relx-generated riak script again to do the rpc.